Update README.md
I added more up-to-date contact information and generally created a README.md which is current and relevant rather than old without much info at all.  It's good to keep your Profile Information including contact methods current so you can communicate with like-minded individuals who are on the same journey as you are.  We need each other to survive, that's the point, that's the fact of the matter.  Nobody can do the real thing alone by themselves.  Lean on somebody and don't be afraid to admit you need help.  It's a sign that you're growing and becoming a better person when you ask a peer for help.